refactor: moved codes into own dir
