+++ src/main/java/io/nova/core/codes/MouseCodes.java
@@ -1,4 +1,4 @@
-package io.nova.core;
+package io.nova.core.codes;
 
 public class MouseCodes {
     public static final int
